refactor(utils): migrate browserCheck to TS
