+++ src/utils/browserCheck.ts
