+++ src/utils/webrtc/index.js
@@ -21,7 +21,7 @@ import BrowserStorage from '../../services/BrowserStorage.js'
 import { getTalkConfig } from '../../services/CapabilitiesManager.ts'
 import { fetchSignalingSettings } from '../../services/signalingService.js'
 import store from '../../store/index.js'
-import { isSafari } from '../browserCheck.js'
+import { isSafari } from '../browserCheck.ts'
 import CancelableRequest from '../cancelableRequest.js'
 import Signaling from '../signaling.js'
 import SignalingTypingHandler from '../SignalingTypingHandler.js'
